backend/pyproject.toml
[project]
name = "massai-backend"
version = "0.1.0"
description = "FastAPI backend for the MaaSAI production monitoring tool."
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14.1",
    "asyncpg>=0.30.0",
    "fastapi>=0.115.12",
    "httpx>=0.28.1",
    "jsonschema>=4.23.0",
    "pydantic-settings>=2.8.1",
    "python-jose[cryptography]>=3.3.0",
    "sqlalchemy[asyncio]>=2.0.39",
    "uvicorn[standard]>=0.34.0",
    "web3>=7.13.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra"
